Update handling of Project EvaluatorSettings
Currently, using pkl-go to evaluate using a PklProject always yields a warning:
```
warn: Cannot find field on Go struct `pkl.ProjectEvaluatorSettings` matching Pkl property `http`. Ensure the Go structs are up to date with Pkl classes either through codegen or manually adding `pkl` tags.
```

The `http` field is populated to avoid the above warning.
I opted to reuse the existing `Http` and `Proxy` structs here, but I'm unsure if this is the preferred approach over defining new structs.
Happy to change direction on this based on feedback.

This PR also changes `ProjectEvaluatorSettings` to `EvaluatorSettings` to reflect that this struct now maps to its own stdlib module.
This is a breaking change, but it matches the one in the stdlib and it's less likely clients are using this API directly.
One concern I have here is that the name `EvaluatorSettings` could potentially cause confusion with `EvaluatorOptions`.
I see value in mirroring the stdlib here, but can revert this if folks feel strongly.