Change object types to be non-pointers during codegen (#136)
Changes the Go code generator to produce non-pointer types for non-nullables.
The problem this aims to solve is to address an issue where pointers are used both to represent values that are nullable, and values that are objects. This causes ambiguity (is this a pointer because it's nullable, or simply because it's an object?).

Users that prefer the old style for whatever reason can still use an older version of pkl-gen-go.
The decoder determines the correct type based on whether the declared type is a pointer or not.

---------

Co-authored-by: Thomas Purchas <tpurchas@gmail.com>